convert c files to cpp files
